server/requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.2
django-cors-headers==3.14.0
django-oauth-toolkit==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
dnspython==2.3.0
drf-social-oauth2==1.3.0
drf-spectacular==0.26.2
drf-yasg==1.21.5
ecdsa==0.18.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
jwcrypto==1.4.2
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
pyrsistent==0.19.3
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
wrapt==1.15.0
